Avoid ~infinite table max logical width inside flex and grid containers.
An infinite (or actually kTableMaxWidth, which is 1000000px) max logical
width doesn't work well inside flexboxes, because
LayoutFlexibleBox::ComputeInnerFlexBaseSizeForChild() will use a child's
MaxPreferredLogicalWidth() as main axis extent if it's auto.

Grid containers have similar issues with such tables, so avoid it for
them as well.

Don't apply this to deprecated flexbox, as that would cause two tests
failures:
* fast/block/block-size-integer-overflow.html
* fast/deprecated-flexbox/box-size-integer-overflow.html

Bug: 850566
Change-Id: I9984459801ac229575e8177a9d683088e723db6c