[LayoutNG] Avoid creating empty lines at bogus (unset) BFC offset 0,0.
This affects auto-positioned out-of-flow nodes that are handled as part
of inline layout, and other things too.

This fix will also make it possible to use the position of the last
child fragment to calculate the block size of a block (which I happen to
be working on).

Cq-Include-Trybots: luci.chromium.try:linux_layout_tests_layout_ng;master.tryserver.chromium.linux:linux_layout_tests_layout_ng
Change-Id: Ie3f99d850de81bd349e89c624f86cce95b38eede